Minor refactor process undergoing
