--- Modules/Info/description.json
@@ -1,4 +1,0 @@
-{
-  "monark": "Gerar um tweet aleat\u00F3rio do Monark",
-  "help": "Kimi comes in clutch with the `!help` command!"
-}
