--- Modules/Roles/GrantRole.cs
@@ -1,21 +1,0 @@
-﻿using Discord;
-using Discord.Commands;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Kimi.Modules.Roles
-{
-    class GrantRole : ModuleBase<SocketCommandContext>
-    {
-        //[Command("brasil")]
-        //public async Task GrantWCRole()
-        //{
-        //    var user = Context.User;
-        //    var role = Context.Guild.Roles.FirstOrDefault(x => x.Name == "RoleName");
-        //    await (user as IGuildUser).AddRoleAsync(role);
-        //}
-    }
-}
