--- PrefixHandler.cs
@@ -1,54 +1,0 @@
-﻿using Discord;
-using Discord.Commands;
-using Discord.WebSocket;
-using System;
-using System.Collections.Generic;
-using Microsoft.Extensions.Configuration;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Reflection;
-
-namespace Kimi
-{
-    public class PrefixHandler
-    {
-        private readonly DiscordSocketClient _client;
-        private readonly CommandService _commands;
-        private readonly IConfigurationRoot _config;
-
-        public PrefixHandler(DiscordSocketClient client, CommandService commands, IConfigurationRoot config)
-        {
-            _client = client;
-            _commands = commands;
-            _config = config;
-        }
-
-        public async Task InitializeAsync()
-        {
-            _client.MessageReceived += HandleCommandAsync;
-        }
-
-        public void AddModule<T>()
-        {
-            _commands.AddModuleAsync<T>(null);
-        }
-
-        private async Task HandleCommandAsync(SocketMessage messageParam)
-        {
-            var message = messageParam as SocketUserMessage;
-            if (message == null) return;
-
-            int argPos = 0;
-
-            if (!(message.HasCharPrefix(_config["Prefix"][0], ref argPos) || message.HasMentionPrefix(_client.CurrentUser, ref argPos)) || message.Author.IsBot) return;
-
-            var context = new SocketCommandContext(_client, message);
-
-            await _commands.ExecuteAsync(
-                context: context,
-                argPos: argPos,
-                services: null);
-        }
-    }
-}
