+++ src/Kimi.Core/PrefixHandler.cs
@@ -1,0 +1,53 @@
+﻿using Discord.Commands;
+using Discord.WebSocket;
+using Microsoft.Extensions.Configuration;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Kimi.Core
+{
+    internal class PrefixHandler
+    {
+        private readonly DiscordSocketClient _client;
+        private readonly CommandService _commands;
+        private readonly IConfigurationRoot _config;
+        private readonly IServiceProvider _services;
+
+        public PrefixHandler(DiscordSocketClient client, CommandService commands, IConfigurationRoot config, IServiceProvider services)
+        {
+            _client = client;
+            _commands = commands;
+            _config = config;
+            _services = services;
+        }
+
+        public async Task InitializeAsync()
+        {
+            await _commands.AddModulesAsync(Assembly.GetEntryAssembly(), _services);
+            _client.MessageReceived += HandleCommandAsync;
+        }
+
+        private async Task HandleCommandAsync(SocketMessage messageParam)
+        {
+            var message = messageParam as SocketUserMessage;
+            if (message == null) return;
+
+            int argPos = 0;
+
+            if (!(message.HasCharPrefix(_config["Prefix"][0], ref argPos) || message.HasMentionPrefix(_client.CurrentUser, ref argPos)) 
+                || message.Author.IsBot)
+                return;
+
+            var context = new SocketCommandContext(_client, message);
+
+            await _commands.ExecuteAsync(
+                context: context,
+                argPos: argPos,
+                services: _services);
+        }
+    }
+}
