+++ src/Kimi.Core/appconfig.json
@@ -1,0 +1,4 @@
+{
+  "Token": "OTg4Mjg1ODcwMTQ0NjMwODM0.GYwUuv.muIUfKqQkGW_0xRlAh4LAm3HLm61gZMpwqqioQ",
+  "Prefix": "!"
+}
