--- vip.json
@@ -1,3 +1,0 @@
-{
-  "AdminID": 191604848423075840
-}
